refactor(markdown): split document section helpers
Move document-level outline, viewer, layer, and structure rendering helpers out of the main Markdown formatter so page rendering can be reviewed separately from document metadata sections.
